+++ cpp/src/common/nvtx.cu
@@ -147,16 +147,7 @@ void PUSH_RANGE(const char *name) {
   nvtxRangePushEx(&eventAttrib);
 }
 
-#ifdef ENABLE_EMPTY_MARKER_KERNEL
-__global__ void emptyMarkerKernel() {}
-#endif  // ENABLE_EMPTY_MARKER_KERNEL
-
-void POP_RANGE() {
-  nvtxRangePop();
-#ifdef ENABLE_EMPTY_MARKER_KERNEL
-  emptyMarkerKernel<<<1, 1>>>();
-#endif
-}
+void POP_RANGE() { nvtxRangePop(); }
 
 #else  // NVTX_ENABLED