+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - PR #2638: Improve cython build with custom `build_ext`
 
 ## Improvements
+- PR #2873: Remove empty marker kernel code for NVTX markers
 - PR #2796: Remove tokens of length 1 by default for text vectorizers
 - PR #2741: Use rapids build packages in conda environments
 - PR #2735: Update seed to random_state in random forest and associated tests